chapter 19, 'Strings':
    - version: 0.1.1: compiler and vm are separate entities.
    - the vm::readObjString() function parses the ObjString entity.

TODO:
    1. proceed from clause 19.4 'Operations on Strings';
    2. implement the memory manager (challenge #3, c. 14);
    3. the run() function shall be optimized;
    4. add stack overflow and underflow control (challenge #3, c. 15);
    5. add support for string interpolation (challenge #1, c. 16);
    6. add support for ternary operator (challenge #3, c. 17);
    7. the types and their representation are the subject of further
       refactoring;
    8. fix compiler's error logger, which points to a wrong lexeme
       when we pass over to the 'op_negate' an incorrect expression of king
       -((false)).

+++ common/values.c
@@ -36,7 +36,7 @@ printValue(i32 value, ValueType type)
 			printf("%d", value);
 		break;
 		case val_obj:
-			printObject((u32)value);
+			printObject(value);
 		break;
 	}
 }
